requirements.txt
mkdocs==1.6.1
mkdocs-material==9.7.1
mkdocs-macros-plugin==1.5.0
python-frontmatter==1.1.0